module: memory: partition out memory APIs from generic.c
Move all the module memory APIs into common and heap specific files with
no other changes.

This is to enable easier abstraction and partitioning around memory for
modules and userspace going forward.
