+++ src/include/sof/audio/module_adapter/module/generic.h
@@ -227,6 +227,9 @@ static inline void *mod_zalloc(struct processing_module *mod, size_t size)
 }
 
 int mod_free(struct processing_module *mod, const void *ptr);
+void module_adapter_mem_free(struct processing_module *mod);
+struct processing_module *module_adapter_mem_alloc(const struct comp_driver *drv,
+						   const struct comp_ipc_config *config);
 #if CONFIG_COMP_BLOB
 struct comp_data_blob_handler *mod_data_blob_handler_new(struct processing_module *mod);
 void mod_data_blob_handler_free(struct processing_module *mod, struct comp_data_blob_handler *dbh);
@@ -235,6 +238,7 @@ void mod_data_blob_handler_free(struct processing_module *mod, struct comp_data_
 const void *mod_fast_get(struct processing_module *mod, const void * const dram_ptr, size_t size);
 void mod_fast_put(struct processing_module *mod, const void *sram_ptr);
 #endif
+void mod_resource_init(struct processing_module *mod);
 void mod_free_all(struct processing_module *mod);
 int module_prepare(struct processing_module *mod,
 		   struct sof_source **sources, int num_of_sources,
