+++ test/cmocka/src/audio/eq_fir/CMakeLists.txt
@@ -24,6 +24,8 @@ add_library(audio_for_eq_fir STATIC
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module_adapter.c
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module_adapter_ipc3.c
 	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/generic.c
+	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/memory-common.c
+	${PROJECT_SOURCE_DIR}/src/audio/module_adapter/module/memory-heap.c
 	${PROJECT_SOURCE_DIR}/src/audio/buffers/comp_buffer.c
 	${PROJECT_SOURCE_DIR}/src/audio/buffers/audio_buffer.c
 	${PROJECT_SOURCE_DIR}/src/audio/source_api_helper.c
